+++ var_hash.h
@@ -17,7 +17,7 @@ var_hash* var_hash_init();
 var_elem* var_hash_add_name ( var_hash**, char* );
 var_elem* var_hash_insert (var_hash** , var_elem* );
 var_elem* var_hash_find (var_hash** , char* );
-int var_hash_size( var_hash** );
+unsigned int var_hash_size( var_hash** );
 char** var_hash_names( var_hash** );
 void var_hash_names_free( char** hash_names, int size );
 void var_hash_free( var_hash** );